+++ Helpers/VersionHelper.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Version Helper
  *
@@ -8,11 +9,13 @@
  * PHP version 8.1
  *
  * @category  Helpers
- * @package   Modules\WhiteLabel\Helpers
+ *
  * @author    Rihards Simanovics <rihards.s@griffin-web.studio>
  * @copyright 2024 Rihards Simanovics
  * @license   GNU GPLv3
+ *
  * @version   GIT: <git_id>
+ *
  * @link      https://invoiceshelf.com
  */
 
@@ -33,9 +36,8 @@ class VersionHelper
      * This method checks the application version against a specified version using a comparison operator.
      * It temporarily aliases the Setting class to access the version information.
      *
-     * @param string $operator The comparison operator (e.g., '<', '>=').
-     * @param string $version  The version to compare against.
-     *
+     * @param  string  $operator  The comparison operator (e.g., '<', '>=').
+     * @param  string  $version  The version to compare against.
      * @return bool True if the comparison is valid, false otherwise.
      */
     public static function checkAppVersion($operator, $version)
@@ -49,12 +51,12 @@ class VersionHelper
     /**
      * Alias old namespace classes to new ones if necessary.
      *
-     * @param string $oldClass The old class name with namespace.
-     * @param string $newClass The new class name with namespace.
+     * @param  string  $oldClass  The old class name with namespace.
+     * @param  string  $newClass  The new class name with namespace.
      */
     public static function aliasClass($newClass, $oldClass)
     {
-        if (!class_exists($oldClass) && class_exists($newClass)) {
+        if (! class_exists($oldClass) && class_exists($newClass)) {
             class_alias($newClass, $oldClass);
         }
     }